refactor: move ops and change classes in internal core to builder package
