+++ cli/flamingock-cli/src/main/java/io/flamingock/cli/service/AuditService.java
@@ -34,8 +34,8 @@ import io.flamingock.internal.common.core.context.Dependency;
 import io.flamingock.internal.common.core.audit.issue.AuditEntryIssue;
 import io.flamingock.internal.common.core.recovery.FixResult;
 import io.flamingock.internal.common.core.recovery.Resolution;
-import io.flamingock.internal.core.builder.ops.OpsClient;
-import io.flamingock.internal.core.builder.ops.OpsClientBuilder;
+import io.flamingock.internal.core.builder.OpsClient;
+import io.flamingock.internal.core.builder.OpsClientBuilder;
 import io.flamingock.internal.core.configuration.core.CoreConfiguration;
 import io.flamingock.internal.core.configuration.community.CommunityConfiguration;
 import io.flamingock.internal.core.context.SimpleContext;
