+++ community/flamingock-community/src/main/java/io/flamingock/community/Flamingock.java
@@ -15,7 +15,7 @@
  */
 package io.flamingock.community;
 
-import io.flamingock.internal.core.builder.change.CommunityChangeRunnerBuilder;
+import io.flamingock.internal.core.builder.CommunityChangeRunnerBuilder;
 import io.flamingock.internal.core.builder.FlamingockFactory;
 
 public final class Flamingock {
