+++ core/target-systems/couchbase-target-system/src/test/java/io/flamingock/targetsystem/couchbase/CouchbaseTargetSystemTest.java
@@ -29,7 +29,7 @@ import io.flamingock.common.test.cloud.prototype.PrototypeStage;
 import io.flamingock.internal.common.core.util.Deserializer;
 import io.flamingock.internal.common.cloud.vo.TargetSystemAuditMarkType;
 import io.flamingock.internal.common.couchbase.CouchbaseCollectionHelper;
-import io.flamingock.internal.core.builder.change.CloudChangeRunnerBuilder;
+import io.flamingock.internal.core.builder.CloudChangeRunnerBuilder;
 import io.flamingock.internal.core.builder.FlamingockFactory;
 import io.flamingock.internal.util.constants.CommunityPersistenceConstants;
 import io.flamingock.internal.core.runner.PipelineExecutionException;
